Cleanup audit rules (#2)
Some of rules in base set do not exist in the version of auditd
we distribute. A few files in STIG rules do not exist in TrueNAS
and prevent auditd load.

This commit also adds an optional prefix argument for script to
generate privileged rules. This is due to script being run outside
chroot environment in scale_build.